complete moving of resolving CTX list from FilterOperation
